kcalzap/requirements.txt
flask==3.0.0
python-dotenv==1.0.0
openai==1.12.0
twilio==8.12.0
stripe==7.11.0
requests==2.31.0
Pillow==10.2.0
